implement more security for api routes

+++ src/routes/taskroutes.js
@@ -5,7 +5,7 @@ router.use(express.json());
 
 
 const Task = require('./../models/task');
-
+const verifyToken = require('./../utils/jwtFunctions');
 router.use(express.urlencoded({ extended: true }));
 
 //conseguir todas las tareas registradas
@@ -31,10 +31,12 @@ router.post('/task:tittle', async (req,res)=>{
 
 //crear una tarea
 router.post('/newtask', async (req, res)=>{
-
+   
     const taskifexist = await Task.findOne({ tittle: req.body.tittle });
-    // console.log(req.params.tittle)
-    var task = {tittle : req.body.tittle, description : req.body.description}
+    
+    var task = {tittle : req.body.tittle, description : req.body.description, userId : req.body.userId}
+
+    // JWT LOGIC
     const authorization = req.get('authorization');
     let token = '';
     if(authorization && authorization.toLocaleLowerCase().startsWith('bearer')){
@@ -48,6 +50,9 @@ router.post('/newtask', async (req, res)=>{
     if(!token || !decodedToken.id){
         return res.status(401).json({error: 'Invalid token or missing authorization'});
     }
+    // END JWT LOGIC
+    
+
     if (taskifexist) {
         res.status(404).send('La tarea ya esta registrada')
     } else {
@@ -61,11 +66,17 @@ router.post('/newtask', async (req, res)=>{
 router.delete('/deletetask', async (req, res) =>{
 
     const taskifexist = await Task.findOne({ tittle: req.body.tittle });
-   
+    
+    
     if (taskifexist) {
-        await Task.findByIdAndDelete(taskifexist.id)
-        res.status(200).send('Tarea eliminada con exito')
-    } else {res.status(500).json('La tarea no existe en la base de datos')}
+        if (taskifexist.userId === req.body.userId) {
+
+            await Task.findByIdAndDelete(taskifexist.id)
+            res.status(200).send('Task deleted')
+
+        } else{res.status(401).json('You can only delete your tasks')}
+
+    } else {res.status(401).json('Task doasent exist')}
 })
 
 
@@ -75,26 +86,25 @@ router.delete('/deletetask', async (req, res) =>{
 router.post('/updatetask', async (req, res ) =>{
 
    
-
     //find task for update
     const tasktoupdate = await Task.findOne({ tittle: req.body.tittle });
 
     //if exists update
     if (tasktoupdate) {
-        await Task.findByIdAndUpdate({ _id : tasktoupdate.id}, { description : req.body.description});
-        // const taskupdated = await Task.findOne({ tittle: req.body.tittle }); 
-        res.status(200).json('Tarea modificada con exito');
+        if (tasktoupdate.userId === req.body.userId) {
+
+            await Task.findByIdAndUpdate({ _id : tasktoupdate.id}, { description : req.body.description});
+            // const taskupdated = await Task.findOne({ tittle: req.body.tittle }); 
+            res.status(200).json('Tarea modificada con exito');
+            
+        } else{res.status(401).json('You can only update your tasks')}
     }
-    res.status(500).json('La tarea no existe en la base de datos');
+    res.status(401).json('Task doasent exist');
 })
 
 
 //mostrar la informacion de una tarea ya sea buscando mediante la descripccion o el titulo
 router.post('/gettask', async ()=>{
-
-
-
-
 
 
 })
